# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PromptPlace API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/inquiries": {
      "post": {
        "summary": "1:1 문의 등록",
        "tags": [
          "Inquiry"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "receiver_id",
                  "type",
                  "title",
                  "content"
                ],
                "properties": {
                  "receiver_id": {
                    "type": "integer",
                    "description": "수신자 회원 ID"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "buyer",
                      "non_buyer"
                    ]
                  },
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "문의 등록 성공"
          }
        }
      }
    },
...
    "/api/members": {
      "patch": {
        "summary": "회원 정보 수정",
        "tags": [
          "Member"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "회원 정보 수정 성공"
          }
        }
      }
    },
...
    "/api/prompts": {
      "post": {
        "summary": "프롬프트 작성",
        "tags": [
          "Prompts"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "prompt",
                  "prompt_result",
                  "description",
                  "price",
                  "tags",
                  "model",
                  "is_free"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "prompt": {
                    "type": "string"
                  },
                  "prompt_result": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "usage_guide": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "model": {
                    "type": "string"
                  },
                  "is_free": {
                    "type": "boolean"
                  },
                  "has_image": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "업로드 성공"
          },
          "400": {
            "description": "필수 필드 누락"
          },
          "401": {
            "description": "인증 실패"
          }
        }
      }
    },
...
    "/api/reports": {
      "post": {
        "summary": "프롬프트 신고 등록",
        "tags": [
          "Report"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "prompt_id",
                  "report_type",
                  "description"
                ],
                "properties": {
                  "prompt_id": {
                    "type": "integer",
                    "description": "신고 대상 프롬프트 ID"
                  },
                  "report_type": {
                    "type": "string",
                    "enum": [
                      "inappropriate",
                      "spam",
                      "etc"
                    ],
                    "description": "신고 유형"
                  },
                  "description": {
                    "type": "string",
                    "description": "신고 상세 설명"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "신고 등록 성공"
          },
          "401": {
            "description": "인증 실패"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "get": {
        "summary": "신고된 프롬프트 목록 조회 (관리자 전용)",
        "tags": [
          "Report"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "페이징 커서"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "가져올 개수"
          }
        ],
        "responses": {
          "200": {
            "description": "신고된 프롬프트 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reports": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "report_id": {
                            "type": "integer"
                          },
                          "report_type": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "is_read": {
                            "type": "boolean"
                          },
                          "prompt": {
                            "type": "object",
                            "properties": {
                              "prompt_id": {
                                "type": "integer"
                              },
                              "title": {
                                "type": "string"
                              }
                            }
                          },
                          "reporter": {
                            "type": "object",
                            "properties": {
                              "user_id": {
                                "type": "integer"
                              },
                              "nickname": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        }
      }
    },
...
    "/api/tips": {
      "get": {
        "summary": "팁 목록 조회",
        "tags": [
          "Tip"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "페이지 번호"
          },
          {
            "in": "query",
            "name": "size",
            "schema": {
              "type": "integer"
            },
            "description": "한 페이지당 항목 수"
          }
        ],
        "responses": {
          "200": {
            "description": "팁 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "tip_id": {
                            "type": "integer"
                          },
                          "title": {
                            "type": "string"
                          },
                          "content": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "팁 생성",
        "tags": [
          "Tip"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "writer_id": {
                    "type": "integer"
                  },
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "팁 생성 성공"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "관리자 권한 없음"
          }
        }
      }
    },
...
  },
...
}